+++ library/observables-connect/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@w3ux/observables-connect",
   "license": "GPL-3.0-only",
-  "version": "0.9.0-alpha.4",
+  "version": "0.9.0-alpha.5",
   "type": "module",
   "scripts": {
     "clear": "rm -rf node_modules dist tsconfig.tsbuildinfo",
