+++ library/observables-connect/src/index.ts
@@ -1,7 +1,7 @@
 /* @license Copyright 2024 w3ux authors & contributors
 SPDX-License-Identifier: GPL-3.0-only */
 
-import type { ExtensionsStatus, ExtensionStatus } from '@w3ux/types'
+import type { ExtensionsStatus } from '@w3ux/types'
 import { BehaviorSubject } from 'rxjs'
 
 // Whether extensions are being checked
@@ -11,22 +11,3 @@ export const checkingExtensions$ = _gettingExtensions.asObservable()
 // Discovered extensions along with their status
 export const _extensionsStatus = new BehaviorSubject<ExtensionsStatus>({})
 export const extensionsStatus$ = _extensionsStatus.asObservable()
-
-// Sets an extension status
-export const setExtensionStatus = (id: string, status: ExtensionStatus) => {
-  const value = _extensionsStatus.getValue()
-  value[id] = status
-  _extensionsStatus.next(value)
-}
-
-// Removes an extension status
-export const removeExtensionStatus = (id: string) => {
-  const { [id]: _, ...rest } = _extensionsStatus.getValue()
-  _extensionsStatus.next(rest)
-}
-
-// Whether an extension can be connected
-export const extensionCanConnect = (id: string) => {
-  const value = _extensionsStatus.getValue()
-  return ![undefined, 'connected'].includes(value[id])
-}
